+++ R/do.radviz.R
@@ -7,15 +7,21 @@
 #' @param x a data.frame or matrix to be projected, with column names matching row names in springs
 #' @param springs a matrix of 2D dimensional anchor coordinates, as returned by \code{\link{make.S}}
 #' @param trans a transformation to be applied to the data before projection
-#' @param label.color the color of springs for visualization
-#' @param label.size the size of labels
+#' @param scaling a scaling factor applied to data before the projection.
 #' @param type character string specifying the method used for obtaining the springs. 
 #' 				Current methods are: Radviz, Freeviz and Graphviz. When not provided, \code{type} is 
 #' 				derived from the other inputs
 #' @param graph \code{igraph} object (only relevant for result obtained from \code{\link{do.optimGraphviz}} analysis) 
+#' @param label.color deprecated, use \code{\link{plot.radviz}} instead
+#' @param label.size deprecated, use \code{\link{plot.radviz}} instead
 #' 
-#' @details The function expects that at least some of the column names in df will be matched
-#'            by row names in springs
+#' @details The function expects that at least some of the column names in x 
+#' will be matched by all row names in springs.
+#' The scaling factor can be used to increase the distance between points, 
+#' making it useful in situations where all points are pulled together either 
+#' because of similar values or large number of channels.
+#' The scaling is applied **after** the transformation by \code{trans}.
+#' The scaling idea is taken from [Artur & Minghim 2019](https://doi.org/10.1016/j.cag.2019.08.015).
 #' 
 #' @return an object of class radviz with the following slots:
 #'			\itemize{
@@ -26,6 +32,7 @@
 #'            			\item \code{rx} and \code{ry} the X and Y coordinates of the radviz projection of \code{x} over \code{springs}
 #'            			\item \code{rvalid} an index of points corresponding to an invalid projection (any \code{rx} or \code{ry} is NA)
 #' 				}
+#' 				\item \code{springs}: the matrix containing the spring coordinates.
 #' 				\item \code{type}: character string specifying the method used for obtaining the springs.
 #' 				\item \code{trans}: the function used to transform the data.
 #'        \item \code{graphEdges}: when the input \code{graph} is provided (for a Graphviz analysis),
@@ -39,7 +46,7 @@
 #' 
 #' @aliases do.radviz do.radviz.default
 #' 
-#' @importFrom ggplot2 ggplot aes_string geom_text xlim ylim coord_equal
+#' @importFrom ggplot2 ggplot aes_string geom_text xlim ylim coord_equal GeomLabel
 #' 
 #' @aliases do.radviz do.radviz.default
 #' @author Yann Abraham
@@ -47,6 +54,7 @@
 do.radviz <- function(x,
                       springs,
                       trans=do.L,
+                      scaling=1,
                       label.color='orangered4',
                       label.size=NA,
                       type=NULL,
@@ -69,7 +77,10 @@ do.radviz <- function(x,
   if(!is.null(trans)) {
     mat <- apply(mat, 2, trans)
   }
-  weights <- mat/matrix(rep(rowSums(mat),each=ncol(mat)),nrow=nrow(mat),byrow=T)
+  weights <- mat/matrix(rep(rowSums(mat^scaling),
+                            each=ncol(mat)),
+                        nrow=nrow(mat),
+                        byrow=T)
   rx <- colSums(t(weights)*springs[,1])
   ry <- colSums(t(weights)*springs[,2])
   rvd <- apply(cbind(rx,ry),1,function(x) any(is.na(x)))
@@ -85,10 +96,10 @@ do.radviz <- function(x,
   
   # automatic check for type of method:
   if(is.null(type))	{
-	  springLengths <-  diag(springs%*%t(springs))
-	  if(any(springLengths < 0.99)){
-		  type <- "Freeviz"
-	  } else type <- "Radviz"
+    springLengths <-  diag(springs%*%t(springs))
+    if(any(springLengths < 0.99)){
+      type <- "Freeviz"
+    } else type <- "Radviz"
   }
   
   # find axis range
@@ -96,23 +107,19 @@ do.radviz <- function(x,
   
   radviz <- list(proj=ggplot(data=x,
                              aes_string(x='rx',y='ry'))+
-                   geom_text(data = data.frame(springs,
-                                               Channel=factor(rownames(springs),
-                                                              levels=rownames(springs))),
-                             aes_string(x='X1',y='X2',label='Channel'),
-                             color=label.color,
-                             size=label.size)+
                    coord_equal()+
                    xlim(lims)+
                    ylim(lims)+
                    theme_radviz(),
-		   		type=type,
-		   		trans=trans)
-		
-		if(!is.null(graph)){
-			radviz$graphEdges <- igraph::as_data_frame(graph)
-			radviz$type <- "Graphviz"
-		}
+                 springs = springs,
+                 type=type,
+                 trans=trans)
+  
+  if(!is.null(graph)){
+    radviz$graphEdges <- igraph::as_data_frame(graph)
+    radviz$type <- "Graphviz"
+  }
+  
   class(radviz) <- 'radviz'
   return(radviz)
 }