+++ R/plot.radviz.R
@@ -7,18 +7,22 @@
 #' @param x a radviz object as produced by \code{\link{do.radviz}}
 #' @param main [Optional] a title to the graph, displayed on top
 #' @param anchors.only by default only plot the anchors so that other methods can easily be chained
-#' @param anchors.filter filter out anchors with low contributions to the projection
+#' @param anchors.filter filter out anchors with low contributions to the projection (superseded)
 #' @param label.color the color of springs for visualization
-#' @param label.size the size of labels
+#' @param label.size the size of the anchors (see \href{https://ggplot2.tidyverse.org/articles/articles/faq-customising.html}{customizing ggplot2} for details on default value)
 #' @param ...	further arguments to be passed to or from other methods (not implemented)
-#' @param point.color deprecated, use \code{\link{geom_point}} instead
-#' @param point.shape deprecated, use \code{\link{geom_point}} instead
-#' @param point.size deprecated, use \code{\link{geom_point}} instead
-#' @param add deprecated, use \code{\link{geom_point}} instead
+#' @param point.color deprecated, use \code{\link[ggplot2]{geom_point}} instead
+#' @param point.shape deprecated, use \code{\link[ggplot2]{geom_point}} instead
+#' @param point.size deprecated, use \code{\link[ggplot2]{geom_point}} instead
+#' @param add deprecated, use \code{\link[ggplot2]{geom_point}} instead
 #' 
-#' @details by default the plot function only shows the anchors. Extra geoms are required to display the data.
-#' When \code{anchors.filter} is a number and type is not Radviz, any springs whose length is lower than this number will be filtered out 
-#' of the visualization. This has no effect on the projection itself. 
+#' @details by default the plot function only shows the anchors. Extra geoms are 
+#' required to display the data.
+#' When \code{anchors.filter} is a number and type is not Radviz, any springs 
+#' whose length is lower than this number will be filtered out 
+#' of the visualization. This has no effect on the projection itself. Please note
+#' that this parameter is being superseded by the \code{\link{anchor.filter}} 
+#' function.
 #' 
 #' @return the internal ggplot2 object, allowing for extra geoms to be added
 #' 
@@ -33,12 +37,12 @@
 #' plot(rv)+geom_point(aes(color=Species))
 #' 
 #' @author Yann Abraham
-#' @importFrom ggplot2 ggtitle geom_point
+#' @importFrom ggplot2 ggtitle geom_point geom_text GeomLabel
 #' @export
 plot.radviz <- function(x,
                         main=NULL,
                         anchors.only=TRUE,
-                        anchors.filter=NULL,
+                        anchors.filter = NULL,
                         label.color=NULL,
                         label.size=NULL,
                         point.color,
@@ -55,16 +59,17 @@ plot.radviz <- function(x,
     warning('point.size is a deprecated argument, use plot(x)+geom_point() and custom aes() to change plot.',call. = FALSE)
   if(!missing(add))
     warning('add is a deprecated argument, use plot(x)+geom_point() and custom aes() to change plot.',call. = FALSE)
+  if(!is.null(anchors.filter)) {
+    warning('anchors.filter is a deprecated argument, use anchor.filter(x)',call. = FALSE)
+    ## apply anchor.filter if argument has been used
+    x <- anchor.filter(x,lim = anchors.filter)
+  }
+  
   ## plot
   p <- x$proj
-  if(!is.null(main)) {
-    p <- p + ggtitle(main)
-  } 
-  if(!anchors.only) { 
-    p <- p + geom_point()
-  }
+  
   if(is.null(label.size)) {
-    label.size <- NA
+    label.size <- GeomLabel$default_aes$size
   }
   if(is.null(label.color)) {
     label.color <- 'orangered4'
@@ -72,22 +77,22 @@ plot.radviz <- function(x,
   if(!is.numeric(label.size)){
     label.size <- as.numeric(label.size)
   }
-  if(!is.null(anchors.filter) & x$type!='Radviz') {
-    r <- rowSums(x$proj$plot_env$springs^2)^0.5
-    df <- p$layers[[1]]$data %>% 
-      subset(r>=anchors.filter)
-  } else if(!is.null(anchors.filter) & x$type=='Radviz') {
-    warning('`anchors.filter` is not relevant for Radviz plots\n')
-    df <- p$layers[[1]]$data
-  } else {
-    df <- p$layers[[1]]$data
+  
+  if(!is.null(main)) {
+    p <- p + ggtitle(main)
   }
-  if(!is.null(label.color) | !is.null(label.size) | !is.null(anchors.filter)) {
-    p$layers[[1]] <- geom_text(data = df,
-                               aes_string(x='X1',y='X2',label='Channel'),
-                               color=label.color,
-                               size=label.size)
+  
+  if(!anchors.only) { 
+    p <- p + geom_point()
   }
+  
+  p <- p+
+    geom_text(data = data.frame(x$springs,
+                                Channel=factor(rownames(x$springs),
+                                               levels=rownames(x$springs))),
+              aes_string(x='X1',y='X2',label='Channel'),
+              color=label.color,
+              size=label.size)
   
   return(p)
 }